Refactor helpers into utils and improve error handling
Moved helper functions from helpers.go to utils.go and deleted helpers.go. Updated function names to use camelCase and improved error handling and logging in effects.go and token_transfer.go. Enhanced trustline processing to generate balance authorization state changes for new trustlines. Renamed helpers_test.go to utils_test.go to match the refactor.
