+++ internal/indexer/processors/utils_test.go
